+++ mayan/apps/templating/classes.py
@@ -61,6 +61,8 @@ class Template:
                     'OPTIONS': {
                         'builtins': [
                             'mathfilters.templatetags.mathfilters',
+                            'mayan.apps.templating.templatetags.templating_datetime_tags',
+                            'mayan.apps.templating.templatetags.templating_regex_tags',
                             'mayan.apps.templating.templatetags.templating_tags'
                         ]
                     }